WIP: server refactoring #63
