+++ akumulid/signal_handler.cpp
@@ -1,4 +1,5 @@
-#include "signals.h"
+#include "signal_handler.h"
+#include <iostream>
 #include <signal.h>
 #include <boost/exception/all.hpp>
 
@@ -17,13 +18,19 @@ void SignalHandler::wait() {
     sigset_t sset = {};
     sigemptyset(&sset);
     for(auto s: signals_) {
-        sigaddset(&sset, s);
+        std::cout << "adding signal " << s << std::endl;
+        if (sigaddset(&sset, s) != 0) {
+            std::cout << "sigaddset error" << std::endl;
+        }
     }
     int signo = 0;
-    if (sigwait(sset, &signo) != 0) {
+    std::cout << "starting to wait for signals" << std::endl;
+    if (sigwait(&sset, &signo) != 0) {
+        std::cout << "sigwait error" << std::endl;
         std::runtime_error error("`sigwait` error");
         BOOST_THROW_EXCEPTION(error);
     }
+    std::cout << "signal received" << std::endl;
     for (auto fn: handlers_) {
         fn();
     }
