+++ akumulid/signal_handler.h
