+++ unittests/CMakeLists.txt
@@ -93,6 +93,7 @@ add_executable(
     test_tcp_server.cpp
     ../akumulid/ingestion_pipeline.cpp
     ../akumulid/tcp_server.cpp
+    ../akumulid/signal_handler.cpp
     ../akumulid/resp.cpp
     ../akumulid/stream.cpp
     ../akumulid/protocolparser.cpp
